propacienta/requirements/base.txt
pytz==2021.3  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==9.0.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.3.0  # https://github.com/evansd/whitenoise
redis==4.1.2  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
celery==5.2.3  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.2.1  # https://github.com/celery/django-celery-beat
flower==1.0.0  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==3.2.11  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.47.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.13.1  # https://github.com/encode/django-rest-framework
django-cors-headers==3.11.0 # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.21.1
drf-yasg==1.20.0
djoser
django-auditlog
